requirements.txt
PySide6
phonenumberslite~=9.0.18
pdfplumber~=0.11.8